youtrack-time-in-status/requirements.txt
requests
pandas
notebook
python-dotenv